src/app/courses/components/error-dialog/error-dialog.component.ts
import { Component, Inject } from '@angular/core';
import { MAT_DIALOG_DATA, MatDialogTitle, MatDialogContent, MatDialogActions, MatDialogClose } from '@angular/material/dialog';

@Component({
  selector: 'app-error-dialog',
  templateUrl: './error-dialog.component.html',
  styleUrls: ['./error-dialog.component.scss'],
  standalone: true,
  imports: [MatDialogTitle, MatDialogContent, MatDialogActions, MatDialogClose]
})
export class ErrorDialogComponent {

  public showBlinkingText = true;

  constructor(
    @Inject(MAT_DIALOG_DATA) public data: string
  ) { }

}
